TeachingWebsite/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-extensions>=3.2.3
